MONGOCRYPT-509 Create serialization API for FLE2IndexedEqualityEncryptedValueTokens (#548)
* MONGOCRYPT-509 Create serialization API for FLE2IndexedEqualityEncryptedValueTokens

* Different unsigned int issue

* Fixed uint issue

+++ src/mc-writer.c
@@ -29,7 +29,7 @@
                   " bytes, got: %" PRIu64,               \
                   writer->parser_name,                   \
                   (writer->len - writer->pos),           \
-                  (write_size));                         \
+                  (uint64_t) (write_size));              \
       return false;                                      \
    }
